fix: share the browser fixture through conftest, rebuild the CSS bundle
Two CI failures on this PR, both mine.

The UI fixture was imported from test_ui.py into the new hash-navigation
module, and an imported fixture is shadowed by the test's own parameter of
the same name — which flake8 reports as F811, correctly, since the import
really is unused at that point. Caught by the failing set this branch's
predecessor introduced (#428), which is a good sign for the rule. The
fixture and the require-browser helper now live in conftest.py, where a
fixture used by two modules belongs; nothing needs to import them.

The SSO failure panel added Tailwind classes (bg-red-50, border-red-300,
their dark variants) that were not in the committed bundle, so the
"tailwind.min.css is stale" gate fired. Rebuilt with npm run css:build.
